Minor full chain refactoring and formatting

+++ mlreco/iotools/collates.py
@@ -89,7 +89,7 @@ def CollateMinkowski(batch):
                 coords = [sample[key][0] for sample in batch]
                 features = [sample[key][1] for sample in batch]
                 coords, features = ME.utils.sparse_collate(coords, features)
-                result[key] = torch.cat([coords.float(), features.float()], dim=1)
+                result[key] = torch.cat([coords.double(), features.double()], dim=1)
             elif isinstance(batch[0][key],np.ndarray) and len(batch[0][key].shape)==1:
                 result[key] = concat( [ concat( [np.expand_dims(sample[key],1),
                                                 np.full(shape=[len(sample[key]),1],fill_value=batch_id,dtype=np.float32)],
